# api/openapi/releasepilot-v1.yaml
openapi: 3.1.0
info:
  title: ReleasePilot Control Plane API
  version: 0.1.0
  description: >-
    ReleasePilot MVP의 릴리스 요청, 승인, Canary 운영 명령,
    정책 및 감사 조회 API 계약이다.
servers:
  - url: /api/v1
    description: Same-origin API
...
paths:
...
  /projects:
    post:
      tags: [Catalog]
      summary: 프로젝트 생성
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: 프로젝트 생성됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
    get:
      tags: [Catalog]
      summary: 프로젝트 목록 조회
      operationId: listProjects
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 프로젝트 목록
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
                  nextCursor:
                    type: [string, 'null']
  /projects/{projectId}/services:
    post:
      tags: [Catalog]
      summary: 서비스 등록
      operationId: createService
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServiceRequest'
      responses:
        '201':
          description: 서비스 등록됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /connections/clusters:
    post:
      tags: [Catalog]
      summary: Kubernetes 클러스터 연결 등록
      operationId: createClusterConnection
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateClusterConnectionRequest'}
      responses:
        '201':
          description: 미검증 클러스터 연결 생성됨
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ClusterConnection'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
    get:
      tags: [Catalog]
      summary: Kubernetes 클러스터 연결 목록
      operationId: listClusterConnections
      responses:
        '200':
          description: 클러스터 연결 목록
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/ClusterConnection'}
  /connections/prometheus:
    post:
      tags: [Catalog]
      summary: Prometheus 연결 등록
      operationId: createPrometheusConnection
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreatePrometheusConnectionRequest'}
      responses:
        '201':
          description: 미검증 Prometheus 연결 생성됨
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PrometheusConnection'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
    get:
      tags: [Catalog]
      summary: Prometheus 연결 목록
      operationId: listPrometheusConnections
      responses:
        '200':
          description: Prometheus 연결 목록
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/PrometheusConnection'}
...
  /releases/{releaseId}:
    get:
      tags: [Releases]
      summary: 릴리스 상세 조회
      operationId: getRelease
      parameters:
        - $ref: '#/components/parameters/ReleaseId'
      responses:
        '200':
          description: 릴리스 상세
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReleaseDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /releases/{releaseId}/approve:
    post:
      tags: [Approvals]
      summary: 릴리스 승인
      operationId: approveRelease
      parameters:
        - $ref: '#/components/parameters/ReleaseId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisionRequest'
      responses:
        '200':
          description: 승인됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Release'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /releases/{releaseId}/reject:
    post:
      tags: [Approvals]
      summary: 릴리스 거부
      operationId: rejectRelease
      parameters:
        - $ref: '#/components/parameters/ReleaseId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisionRequest'
      responses:
        '200':
          description: 거부됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Release'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: RELEASEPILOT_SESSION
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    ReleaseId:
      name: releaseId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 같은 사용자와 operation에서 24시간 동안 중복 실행을 방지한다.
      schema:
        type: string
        minLength: 16
        maxLength: 128
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    BadRequest:
      description: 요청 형식이 잘못됨
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: 인증 필요
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: 권한 부족
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: 현재 상태 또는 동시 실행과 충돌
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    CreateProjectRequest:
      type: object
      additionalProperties: false
      required: [key, name]
      properties:
        key:
          type: string
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
          maxLength: 63
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 500
    Project:
      type: object
      required: [id, key, name, status]
      properties:
        id:
          type: string
          format: uuid
        key:
          type: string
        name:
          type: string
        description:
          type: [string, 'null']
        status:
          type: string
          enum: [ACTIVE, DISABLED]
        createdAt:
          type: string
          format: date-time
...
    CreateServiceRequest:
      type: object
      additionalProperties: false
      required: [key, name, repositoryUrl, owner]
      properties:
        key:
          type: string
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
          maxLength: 63
        name:
          type: string
          minLength: 1
          maxLength: 100
        repositoryUrl:
          type: string
          format: uri
          pattern: '^https://'
        owner:
          type: string
          minLength: 1
          maxLength: 100
    Service:
      type: object
      required: [id, projectId, key, name, repositoryUrl, owner, status]
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        key:
          type: string
        name:
          type: string
        repositoryUrl:
          type: string
          format: uri
        owner:
          type: string
        status:
          type: string
          enum: [ACTIVE, DISABLED]
        createdAt:
          type: string
          format: date-time
    CreateClusterConnectionRequest:
      type: object
      additionalProperties: false
      required: [name, apiServer, allowedNamespaces, secretRef]
      properties:
        name: {type: string, minLength: 1, maxLength: 100}
        apiServer: {type: string, format: uri, pattern: '^https://'}
        allowedNamespaces:
          type: array
          minItems: 1
          items: {type: string}
        secretRef: {type: string, minLength: 1, maxLength: 255}
    ClusterConnection:
      allOf:
        - $ref: '#/components/schemas/CreateClusterConnectionRequest'
        - type: object
          required: [id, status, createdAt]
          properties:
            id: {type: string, format: uuid}
            status: {type: string, enum: [UNVERIFIED, ACTIVE, INVALID, DISABLED]}
            lastValidatedAt: {type: [string, 'null'], format: date-time}
            createdAt: {type: string, format: date-time}
    CreatePrometheusConnectionRequest:
      type: object
      additionalProperties: false
      required: [name, baseUrl, queryTimeoutSeconds]
      properties:
        name: {type: string, minLength: 1, maxLength: 100}
        baseUrl: {type: string, format: uri}
        secretRef: {type: [string, 'null'], maxLength: 255}
        queryTimeoutSeconds: {type: integer, minimum: 1, maximum: 120}
    PrometheusConnection:
      allOf:
        - $ref: '#/components/schemas/CreatePrometheusConnectionRequest'
        - type: object
          required: [id, status, createdAt]
          properties:
            id: {type: string, format: uuid}
            status: {type: string, enum: [UNVERIFIED, ACTIVE, INVALID, DISABLED]}
            lastValidatedAt: {type: [string, 'null'], format: date-time}
            createdAt: {type: string, format: date-time}
...
    Release:
      type: object
      required:
        - id
        - serviceId
        - environmentId
        - version
        - imageDigest
        - status
        - requestedBy
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        serviceId:
          type: string
          format: uuid
        environmentId:
          type: string
          format: uuid
        version:
          type: string
        imageDigest:
          type: string
        status:
          $ref: '#/components/schemas/ReleaseStatus'
        requestedBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        approvedAt:
          type: [string, 'null']
          format: date-time
        startedAt:
          type: [string, 'null']
          format: date-time
        finishedAt:
          type: [string, 'null']
          format: date-time
    ReleaseDetail:
      allOf:
        - $ref: '#/components/schemas/Release'
        - type: object
          required: [changeSummary, commitSha, pipelineUrl, steps]
          properties:
            changeSummary:
              type: string
            commitSha:
              type: string
            pipelineUrl:
              type: string
              format: uri
            policySnapshot:
              $ref: '#/components/schemas/PolicySnapshotSummary'
            approval:
              anyOf:
                - $ref: '#/components/schemas/ApprovalDecision'
                - type: 'null'
            steps:
              type: array
              items:
                $ref: '#/components/schemas/RolloutStep'
...
    ReleaseStatus:
      type: string
      enum:
        - DRAFT
        - VALIDATING
        - VALIDATION_FAILED
        - PENDING_APPROVAL
        - APPROVED
        - REJECTED
        - CANCELED
        - DEPLOYING
        - ANALYZING
        - PROMOTING
        - PAUSED
        - ROLLING_BACK
        - ROLLED_BACK
        - ROLLBACK_FAILED
        - SUCCEEDED
        - FAILED
    DecisionRequest:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 1000
...
    ApprovalDecision:
      type: object
      required: [decision, decidedBy, reason, decidedAt]
      properties:
        decision:
          type: string
          enum: [APPROVED, REJECTED]
        decidedBy:
          type: string
          format: uuid
        reason:
          type: string
        decidedAt:
          type: string
          format: date-time
    RolloutStep:
      type: object
      required: [index, weight, minimumObservationSeconds, status]
      properties:
        index:
          type: integer
          minimum: 0
        weight:
          type: integer
          minimum: 1
          maximum: 100
        minimumObservationSeconds:
          type: integer
        status:
          type: string
          enum: [PENDING, DEPLOYING, OBSERVING, EVALUATING, PASSED, FAILED, INCONCLUSIVE]
        decision:
          type: [string, 'null']
          enum: [PASS, FAIL, INCONCLUSIVE, null]
    PolicySnapshotSummary:
      type: object
      required: [sourcePolicyVersionId, name, checksum]
      properties:
        sourcePolicyVersionId:
          type: string
          format: uuid
        name:
          type: string
        checksum:
          type: string
...
    Problem:
      type: object
      description: RFC 9457 Problem Details
      required: [type, title, status, code, traceId]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
        traceId:
          type: string
        violations:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
              message:
                type: string
security:
  - sessionCookie: []