# apps/control-plane/src/main/java/kr/releasepilot/controlplane/approval/ApprovalController.java
package kr.releasepilot.controlplane.approval;
import jakarta.validation.Valid;import jakarta.validation.constraints.*;import kr.releasepilot.controlplane.identity.UserAccountPrincipal;import kr.releasepilot.controlplane.release.Release;import org.springframework.security.core.Authentication;import org.springframework.security.core.annotation.AuthenticationPrincipal;import org.springframework.web.bind.annotation.*;import java.time.Instant;import java.util.UUID;
@RestController @RequestMapping("/api/v1/releases/{releaseId}")public class ApprovalController{
 private final ApprovalService service;private final kr.releasepilot.controlplane.release.ReleaseArtifactRepository artifacts;public ApprovalController(ApprovalService service,kr.releasepilot.controlplane.release.ReleaseArtifactRepository artifacts){this.service=service;this.artifacts=artifacts;}
 @PostMapping("/approve")ReleaseResponse approve(@PathVariable UUID releaseId,@RequestHeader("Idempotency-Key")@Size(min=16,max=128)String key,@Valid@RequestBody DecisionRequest r,@AuthenticationPrincipal UserAccountPrincipal actor,Authentication auth){return response(service.decide(releaseId,ApprovalDecision.Decision.APPROVED,r.reason(),key,actor,auth));}
 @PostMapping("/reject")ReleaseResponse reject(@PathVariable UUID releaseId,@RequestHeader("Idempotency-Key")@Size(min=16,max=128)String key,@Valid@RequestBody DecisionRequest r,@AuthenticationPrincipal UserAccountPrincipal actor,Authentication auth){return response(service.decide(releaseId,ApprovalDecision.Decision.REJECTED,r.reason(),key,actor,auth));}
 private ReleaseResponse response(Release r){String digest=artifacts.findByReleaseId(r.getId()).orElseThrow().getImageDigest();return new ReleaseResponse(r.getId(),r.getServiceId(),r.getEnvironmentId(),r.getVersion(),digest,r.getStatus().name(),r.getRequestedBy(),r.getCreatedAt(),r.getApprovedAt(),r.getFinishedAt());}
 public record DecisionRequest(@NotBlank@Size(max=1000)String reason){}public record ReleaseResponse(UUID id,UUID serviceId,UUID environmentId,String version,String imageDigest,String status,UUID requestedBy,Instant createdAt,Instant approvedAt,Instant finishedAt){}
}